requirements_tf214.txt
tensorflow==2.14.0
numpy==1.24.3
scikit-learn==1.3.2
Pillow==10.0.1
